+++ backend/tests/baserow/contrib/database/field/test_field_types.py
@@ -1,19 +1,15 @@
 import pytest
 import json
-from pytz import timezone
-from datetime import date
 from faker import Faker
 from decimal import Decimal
 
 from django.core.exceptions import ValidationError
-from django.utils.timezone import make_aware, datetime
 
 from baserow.core.user_files.exceptions import (
     InvalidUserFileNameError, UserFileDoesNotExist
 )
-from baserow.contrib.database.fields.field_types import DateFieldType
 from baserow.contrib.database.fields.models import (
-    LongTextField, URLField, DateField, EmailField, FileField
+    LongTextField, URLField, EmailField, FileField
 )
 from baserow.contrib.database.fields.handler import FieldHandler
 from baserow.contrib.database.rows.handler import RowHandler
@@ -286,118 +282,6 @@ def test_url_field_type(data_fixture):
 
     field_handler.delete_field(user=user, field=field_2)
     assert len(URLField.objects.all()) == 2
-
-
-@pytest.mark.django_db
-def test_date_field_type_prepare_value(data_fixture):
-    d = DateFieldType()
-
-    f = data_fixture.create_date_field(date_include_time=True)
-    amsterdam = timezone('Europe/Amsterdam')
-    utc = timezone('UTC')
-    expected_date = make_aware(datetime(2020, 4, 10, 0, 0, 0), utc)
-    expected_datetime = make_aware(datetime(2020, 4, 10, 12, 30, 30), utc)
-
-    with pytest.raises(ValidationError):
-        assert d.prepare_value_for_db(f, 'TEST')
-
-    assert d.prepare_value_for_db(f, None) is None
-
-    unprepared_datetime = make_aware(datetime(2020, 4, 10, 14, 30, 30), amsterdam)
-    assert d.prepare_value_for_db(f, unprepared_datetime) == expected_datetime
-
-    unprepared_datetime = make_aware(datetime(2020, 4, 10, 12, 30, 30), utc)
-    assert d.prepare_value_for_db(f, unprepared_datetime) == expected_datetime
-
-    unprepared_datetime = datetime(2020, 4, 10, 12, 30, 30)
-    assert d.prepare_value_for_db(f, unprepared_datetime) == expected_datetime
-
-    unprepared_date = date(2020, 4, 10)
-    assert d.prepare_value_for_db(f, unprepared_date) == expected_date
-
-    assert d.prepare_value_for_db(f, '2020-04-10') == expected_date
-    assert d.prepare_value_for_db(f, '2020-04-11') != expected_date
-    assert d.prepare_value_for_db(f, '2020-04-10 12:30:30') == expected_datetime
-    assert d.prepare_value_for_db(f, '2020-04-10 00:30:30 PM') == expected_datetime
-
-    f = data_fixture.create_date_field(date_include_time=False)
-    expected_date = date(2020, 4, 10)
-
-    unprepared_datetime = datetime(2020, 4, 10, 14, 30, 30)
-    assert d.prepare_value_for_db(f, unprepared_datetime) == expected_date
-
-    unprepared_datetime = make_aware(datetime(2020, 4, 10, 14, 30, 30), amsterdam)
-    assert d.prepare_value_for_db(f, unprepared_datetime) == expected_date
-
-    assert d.prepare_value_for_db(f, '2020-04-10') == expected_date
-    assert d.prepare_value_for_db(f, '2020-04-11') != expected_date
-    assert d.prepare_value_for_db(f, '2020-04-10 12:30:30') == expected_date
-    assert d.prepare_value_for_db(f, '2020-04-10 00:30:30 PM') == expected_date
-
-
-@pytest.mark.django_db
-def test_date_field_type(data_fixture):
-    user = data_fixture.create_user()
-    table = data_fixture.create_database_table(user=user)
-
-    field_handler = FieldHandler()
-    row_handler = RowHandler()
-
-    amsterdam = timezone('Europe/Amsterdam')
-    utc = timezone('utc')
-
-    date_field_1 = field_handler.create_field(user=user, table=table, type_name='date',
-                                              name='Date')
-    date_field_2 = field_handler.create_field(user=user, table=table, type_name='date',
-                                              name='Datetime', date_include_time=True)
-
-    assert date_field_1.date_include_time is False
-    assert date_field_2.date_include_time is True
-    assert len(DateField.objects.all()) == 2
-
-    model = table.get_model(attribute_names=True)
-
-    row = row_handler.create_row(user=user, table=table, values={}, model=model)
-    assert row.date is None
-    assert row.datetime is None
-
-    row = row_handler.create_row(user=user, table=table, values={
-        'date': '2020-4-1',
-        'datetime': '2020-4-1 12:30:30'
-    }, model=model)
-    row.refresh_from_db()
-    assert row.date == date(2020, 4, 1)
-    assert row.datetime == datetime(2020, 4, 1, 12, 30, 30, tzinfo=utc)
-
-    row = row_handler.create_row(user=user, table=table, values={
-        'datetime': make_aware(datetime(2020, 4, 1, 12, 30, 30), amsterdam)
-    }, model=model)
-    row.refresh_from_db()
-    assert row.date is None
-    assert row.datetime == datetime(2020, 4, 1, 10, 30, 30, tzinfo=timezone('UTC'))
-
-    date_field_1 = field_handler.update_field(user=user, field=date_field_1,
-                                              date_include_time=True)
-    date_field_2 = field_handler.update_field(user=user, field=date_field_2,
-                                              date_include_time=False)
-
-    assert date_field_1.date_include_time is True
-    assert date_field_2.date_include_time is False
-
-    model = table.get_model(attribute_names=True)
-    rows = model.objects.all()
-
-    assert rows[0].date is None
-    assert rows[0].datetime is None
-    assert rows[1].date == datetime(2020, 4, 1, tzinfo=timezone('UTC'))
-    assert rows[1].datetime == date(2020, 4, 1)
-    assert rows[2].date is None
-    assert rows[2].datetime == date(2020, 4, 1)
-
-    field_handler.delete_field(user=user, field=date_field_1)
-    field_handler.delete_field(user=user, field=date_field_2)
-
-    assert len(DateField.objects.all()) == 0
 
 
 @pytest.mark.django_db